Fix matplotlib expects ints not floats

+++ aslam_offline_calibration/kalibr/python/kalibr_camera_calibration/CameraCalibrator.py
@@ -729,8 +729,8 @@ def plotOutlierCorners(cself, removedOutlierCorners, fno=1, clearFigure=True, ti
         corners=np.array(corners)
         
         #plot
-        subplot_rows = np.ceil( np.sqrt(len(cself.cameras)) )
-        subplot_cols = np.ceil( np.sqrt(len(cself.cameras)) )
+        subplot_rows = int(np.ceil( np.sqrt(len(cself.cameras)) ))
+        subplot_cols = int(np.ceil( np.sqrt(len(cself.cameras)) ))
         pl.subplot(subplot_rows, subplot_cols, cidx+1)
         pl.title("cam{0}".format(cidx))
